requirements.txt
requests
python-whois
dnspython
certifi
urllib3
